fixed core data

+++ graphgame/Library/Level/Level.swift
@@ -1,7 +1,7 @@
 final class Level: Identifiable {
-	let id: Int
+	let id: Int16
 	
-	init(id: Int) {
+	init(id: Int16) {
 		self.id = id
 	}
 }
